Fix price precision loss for large-value stocks above Number.MAX_SAFE_INTEGER
- OhlcCandle: change OHLC fields from double to long to preserve integer
  price accuracy up to 9.2해 (long max ~9.2×10^18)
- CandleService: replace all doubleValue() calls with longValue() throughout
  buildOneMin, aggregate, fillGaps, computeCurrentBuckets, restoreCurrentBucket,
  applyPrice, flat, tick, and getCandles signatures
- CandleController: pass fallbackPrice as long instead of double
- Stock: add @JsonFormat(STRING) to currentPrice and basePrice so REST API
  serializes them as JSON strings, preventing JS Number precision loss
- TradeEngine: serialize price as toPlainString() in /topic/prices and
  /topic/trades STOMP broadcasts so JS receives exact string values
- useStocks: add explicit Number() conversion for price and basePrice fields
  since backend now sends them as strings

+++ backend/src/main/java/com/spotchzxk/presentation/controller/CandleController.java
@@ -33,7 +33,7 @@ public class CandleController {
         long listedAtMs = stock != null && stock.getListedAt() != null
                 ? stock.getListedAt().atZone(ZoneId.systemDefault()).toInstant().toEpochMilli()
                 : 0L;
-        double fallbackPrice = stock != null ? stock.getCurrentPrice().doubleValue() : 1000.0;
+        long fallbackPrice = stock != null ? stock.getCurrentPrice().longValue() : 1000L;
 
         long beforeMs = before != null ? before : System.currentTimeMillis();
         List<OhlcCandle> candles = candleService.getCandles(stockId, interval, count, beforeMs, listedAtMs, fallbackPrice);
